Some nasty work size computation

+++ program.cl
@@ -1,9 +1,6 @@
 kernel void
-add(const global float* augend,
-    const global float* addend,
-    global float* sum,
-    const unsigned int count) {
-    int id = get_global_id(0);
+add(const global float* augend, const global float* addend, global float* sum, const uint count) {
+    size_t id = get_global_id(0);
 
     if (id < count) {
         sum[id] = augend[id] + addend[id];
